+++ Halforbit.DataStores/FileStores/Serialization/Delimited/Delimiter.cs
@@ -1,0 +1,12 @@
+﻿
+namespace Halforbit.DataStores
+{
+    public static class Delimiter
+    {
+        public const string Tab = "\t";
+
+        public const string Comma = ",";
+
+        public const string Pipe = "|";
+    }
+}